functions/source/KubeManifest/requirements.txt
ruamel.yaml
requests
